+++ Branch-SDK/src/main/java/io/branch/referral/Defines.java
@@ -44,7 +44,6 @@ public class Defines {
         Event("event"),
         Metadata("metadata"),
         CommerceData("commerce_data"),
-        @Deprecated ReferralCode("referral_code"),
         Total("total"),
         Unique("unique"),
         Length("length"),
@@ -125,6 +124,8 @@ public class Defines {
         Params("params"),
         SharedLink("$shared_link"),
         ShareError("$share_error"),
+        SharedChannel("$shared_channel"),
+
         URLSource("android"),
         
         External_Intent_URI("external_intent_uri"),
@@ -249,14 +250,10 @@ public class Defines {
      * </p>
      */
     public enum RequestPath {
-        @Deprecated RedeemRewards("v1/redeem"),
         GetURL("v1/url"),
         GetApp("v1/app-link-settings"),
         RegisterInstall("v1/install"),
-        RegisterClose("v1/close"),
         RegisterOpen("v1/open"),
-        @Deprecated GetCredits("v1/credits/"),
-        @Deprecated GetCreditHistory("v1/credithistory"),
         CompletedAction("v1/event"),
         IdentifyUser("v1/profile"),
         ContentEvent("v1/content-events"),