+++ Branch-SDK/src/main/java/io/branch/referral/SystemObserver.java
@@ -31,13 +31,18 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Locale;
 import java.util.UUID;
+import java.util.concurrent.ScheduledThreadPoolExecutor;
+import java.util.concurrent.TimeUnit;
 
 import static android.content.Context.UI_MODE_SERVICE;
 
 import com.google.android.gms.ads.identifier.AdvertisingIdClient;
 
-import io.branch.coroutines.DataCoroutinesKt;
+import io.branch.coroutines.AdvertisingIdsKt;
+import io.branch.coroutines.InstallReferrersKt;
+import io.branch.data.InstallReferrerResult;
 import io.branch.referral.util.DependencyUtilsKt;
+import kotlin.Pair;
 import kotlin.coroutines.Continuation;
 import kotlin.coroutines.CoroutineContext;
 import kotlin.coroutines.EmptyCoroutineContext;
@@ -444,25 +449,21 @@ abstract class SystemObserver {
      * @param callback {@link AdsParamsFetchEvents} instance to notify process completion
      * @return {@link Boolean} with true if GAID fetch process started.
      */
-    boolean prefetchAdsParams(Context context, AdsParamsFetchEvents callback) {
-        AIDInitializationSessionID_ = PrefHelper.getInstance(context).getSessionID();
-        boolean isPrefetchStarted = false;
+    public void fetchAdId(Context context, AdsParamsFetchEvents callback) {
         if (isFireOSDevice()) {
             setFireAdId(context, callback);
-        } else {
-            isPrefetchStarted = true;
-            if (isHuaweiMobileServicesAvailable(context)) {
-                new HuaweiOAIDFetchTask(context, callback).executeTask();
-            } else {
-                this.executeAdInfoCoroutine(context, callback);
-            }
         }
-        return isPrefetchStarted;
+        else if (isHuaweiMobileServicesAvailable(context)) {
+            this.fetchHuaweiAdId(context, callback);
+        }
+        else {
+            this.fetchGoogleAdId(context, callback);
+        }
     }
 
-    private void executeAdInfoCoroutine(Context context, AdsParamsFetchEvents callback) {
-        if(DependencyUtilsKt.classExists(DependencyUtilsKt.playStoreAdvertisingIdClientClass)) {
-            DataCoroutinesKt.getAdvertisingInfoObject(context, new Continuation<AdvertisingIdClient.Info>() {
+    private void fetchHuaweiAdId(Context context, AdsParamsFetchEvents callback) {
+        if(DependencyUtilsKt.classExists(DependencyUtilsKt.huaweiAdvertisingIdClientClass)) {
+            AdvertisingIdsKt.getHuaweiAdvertisingInfoObject(context, new Continuation<com.huawei.hms.ads.identifier.AdvertisingIdClient.Info>() {
                 @NonNull
                 @Override
                 public CoroutineContext getContext() {
@@ -473,26 +474,17 @@ abstract class SystemObserver {
                 public void resumeWith(Object o) {
                     if (o != null) {
                         try {
-                            AdvertisingIdClient.Info info = (AdvertisingIdClient.Info) o;
+                            com.huawei.hms.ads.identifier.AdvertisingIdClient.Info info = (com.huawei.hms.ads.identifier.AdvertisingIdClient.Info) o;
 
-                            DeviceInfo di = DeviceInfo.getInstance();
-                            if (di == null) {
-                                di = new DeviceInfo(context);
+                            boolean lat = info.isLimitAdTrackingEnabled();
+                            String aid = null;
+
+                            if(!lat) {
+                                aid = info.getId();
                             }
 
-                            SystemObserver so = di.getSystemObserver();
-                            if (so != null) {
-                                boolean latEnabled = info.isLimitAdTrackingEnabled();
-                                so.setLAT(latEnabled ? 1 : 0);
-
-                                // if limit ad tracking is enabled, null any advertising id
-                                if (latEnabled) {
-                                    so.setGAID(null);
-                                }
-                                else {
-                                    so.setGAID(info.getId());
-                                }
-                            }
+                            setLAT(lat ? 1 : 0);
+                            setGAID(aid);
                         }
                         catch (Exception e) {
                             PrefHelper.Debug("Error in continuation: " + e);
@@ -511,26 +503,120 @@ abstract class SystemObserver {
                 callback.onAdsParamsFetchFinished();
             }
 
-            PrefHelper.Debug("Play Store service not found. " +
-                    "If not expected, import com.google.android.gms:play-services-ads-identifier into your gradle dependencies");
+            PrefHelper.Debug("Huawei advertising service not found. " +
+                    "If not expected, import " + DependencyUtilsKt.huaweiAdvertisingIdClientClass + " into your gradle dependencies");
+        }
+    }
+
+
+
+    private void fetchGoogleAdId(Context context, AdsParamsFetchEvents callback) {
+        if(DependencyUtilsKt.classExists(DependencyUtilsKt.playStoreAdvertisingIdClientClass)) {
+            AdvertisingIdsKt.getGoogleAdvertisingInfoObject(context, new Continuation<AdvertisingIdClient.Info>() {
+                @NonNull
+                @Override
+                public CoroutineContext getContext() {
+                    return EmptyCoroutineContext.INSTANCE;
+                }
+
+                @Override
+                public void resumeWith(Object o) {
+                    if (o != null) {
+                        try {
+                            AdvertisingIdClient.Info info = (AdvertisingIdClient.Info) o;
+
+                            boolean lat = info.isLimitAdTrackingEnabled();
+                            String aid = null;
+
+                            if(!lat) {
+                                aid = info.getId();
+                            }
+
+                            setLAT(lat ? 1 : 0);
+                            setGAID(aid);
+                        }
+                        catch (Exception e) {
+                            PrefHelper.Debug("Error in continuation: " + e);
+                        }
+                        finally {
+                            if (callback != null) {
+                                callback.onAdsParamsFetchFinished();
+                            }
+                        }
+                    }
+                }
+            });
+        }
+        else {
+            if (callback != null) {
+                callback.onAdsParamsFetchFinished();
+            }
+
+            PrefHelper.Debug("Play Store advertising service not found. " +
+                    "If not expected, import " + DependencyUtilsKt.playStoreAdvertisingIdClientClass + " into your gradle dependencies");
         }
     }
 
     private void setFireAdId(Context context, AdsParamsFetchEvents callback) {
-        if (context != null) {
-            try {
-                ContentResolver cr = context.getContentResolver();
-                setLAT(Secure.getInt(cr, "limit_ad_tracking"));
-                setGAID(Secure.getString(cr, "advertising_id"));
-                // Don't save advertising id if it's empty/all zeroes/lat=true
-                if (TextUtils.isEmpty(GAIDString_) || GAIDString_.equals(UUID_EMPTY) || LATVal_ == 1) {
-                    setGAID(null);
-                }
-            } catch (Settings.SettingNotFoundException ignored) {}
-        }
+        PrefHelper.Debug("setFireAdId");
+        AdvertisingIdsKt.getAmazonFireAdvertisingInfoObject(context, new Continuation<Pair<? extends Integer, ? extends String>>() {
+            @NonNull
+            @Override
+            public CoroutineContext getContext() {
+                return EmptyCoroutineContext.INSTANCE;
+            }
 
-        if (callback != null) {
-            callback.onAdsParamsFetchFinished();
+            @Override
+            public void resumeWith(@NonNull Object o) {
+                try {
+                    if (o != null) {
+                        Pair<Integer, String> info = (Pair<Integer, String>) o;
+                        setLAT(info.component1());
+                        if (info.component1() == 0) {
+                            setGAID(info.component2());
+                        }
+                        else {
+                            setGAID(info.component2());
+                        }
+                    }
+                }
+                catch (Exception e){
+                    PrefHelper.Debug("Error in continuation: " + e);
+                }
+                finally {
+                    if (callback != null) {
+                        callback.onAdsParamsFetchFinished();
+                    }
+                }
+            }
+        });
+    }
+
+    public void fetchInstallReferrer(Context context_, InstallReferrerFetchEvents callback) {
+        try {
+            InstallReferrersKt.fetchLatestInstallReferrer(context_, new Continuation<InstallReferrerResult>() {
+                @NonNull
+                @Override
+                public CoroutineContext getContext() {
+                    return EmptyCoroutineContext.INSTANCE;
+                }
+
+                @Override
+                public void resumeWith(@NonNull Object o) {
+                    if (o != null) {
+                        InstallReferrerResult latestReferrer = (InstallReferrerResult) o;
+                        AppStoreReferrer.processReferrerInfo(context_, latestReferrer.getLatestRawReferrer(), latestReferrer.getLatestClickTimestamp(), latestReferrer.getLatestInstallTimestamp(), latestReferrer.getAppStore());
+                    }
+                }
+            });
+        }
+        catch(Exception e){
+            PrefHelper.Debug(e.getMessage());
+        }
+        finally {
+            if(callback != null){
+                callback.onInstallReferrersFinished();
+            }
         }
     }
 
@@ -538,6 +624,9 @@ abstract class SystemObserver {
         void onAdsParamsFetchFinished();
     }
 
+    interface InstallReferrerFetchEvents {
+        void onInstallReferrersFinished();
+    }
     /**
      * Get IP address from first non local net Interface
      */
@@ -710,9 +799,5 @@ abstract class SystemObserver {
 
     void setLAT(int lat) {
         LATVal_ = lat;
-    }
-
-    String getAIDInitializationSessionID() {
-        return AIDInitializationSessionID_;
     }
 }